requirements.txt
transformers
datasets
accelerate
peft
trl
bitsandbytes
sentencepiece
faiss-cpu
sentence-transformers
PyYAML

